Remove Snap due to code highlight issue
