+++ docs/data/charts/bars/ColorScaleNoSnap.js
@@ -8,13 +8,9 @@ import HighlightedCode from 'docs/src/modules/components/HighlightedCode';
 
 const series = [{ data: [-2, -9, 12, 11, 6, -4] }];
 
-export default function ColorScale() {
-  const [colorX, setColorX] = React.useState<
-    'None' | 'piecewise' | 'continuous' | 'ordinal'
-  >('piecewise');
-  const [colorY, setColorY] = React.useState<'None' | 'piecewise' | 'continuous'>(
-    'None',
-  );
+export default function ColorScaleNoSnap() {
+  const [colorX, setColorX] = React.useState('piecewise');
+  const [colorY, setColorY] = React.useState('None');
 
   return (
     <Stack direction="column" spacing={1} sx={{ width: '100%', maxWidth: 600 }}>
@@ -24,11 +20,7 @@ export default function ColorScale() {
           sx={{ minWidth: 150 }}
           label="x-axis colorMap"
           value={colorX}
-          onChange={(event) =>
-            setColorX(
-              event.target.value as 'None' | 'piecewise' | 'continuous' | 'ordinal',
-            )
-          }
+          onChange={(event) => setColorX(event.target.value)}
         >
           <MenuItem value="None">None</MenuItem>
           <MenuItem value="piecewise">piecewise</MenuItem>
@@ -40,9 +32,7 @@ export default function ColorScale() {
           sx={{ minWidth: 150 }}
           label="y-axis colorMap"
           value={colorY}
-          onChange={(event) =>
-            setColorY(event.target.value as 'None' | 'piecewise' | 'continuous')
-          }
+          onChange={(event) => setColorY(event.target.value)}
         >
           <MenuItem value="None">None</MenuItem>
           <MenuItem value="piecewise">piecewise</MenuItem>
@@ -151,14 +141,12 @@ export default function ColorScale() {
                 '  }]}',
               ]
             : []),
-
           // ColorY
           ...(colorY === 'None' ? ['  yAxis={[{}]}'] : []),
           ...(colorY === 'continuous'
             ? [
                 '  yAxis={[',
                 `    {`,
-
                 `      type: 'continuous',`,
                 `      min: -10,`,
                 `      max: 10,`,
@@ -173,7 +161,6 @@ export default function ColorScale() {
                 `    type: 'piecewise',`,
                 `    thresholds: [0],`,
                 `    colors: ['red', 'green'],`,
-
                 '  }]}',
               ]
             : []),
