+++ docs/data/charts/bars/ColorScaleNoSnap.tsx
@@ -1,17 +1,19 @@
 import * as React from 'react';
-import { LineChart } from '@mui/x-charts/LineChart';
+import { BarChart } from '@mui/x-charts/BarChart';
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import MenuItem from '@mui/material/MenuItem';
 // @ts-ignore
 import HighlightedCode from 'docs/src/modules/components/HighlightedCode';
 
-export default function ColorScale() {
+const series = [{ data: [-2, -9, 12, 11, 6, -4] }];
+
+export default function ColorScaleNoSnap() {
   const [colorX, setColorX] = React.useState<
     'None' | 'piecewise' | 'continuous' | 'ordinal'
-  >('None');
+  >('piecewise');
   const [colorY, setColorY] = React.useState<'None' | 'piecewise' | 'continuous'>(
-    'piecewise',
+    'None',
   );
 
   return (
@@ -23,12 +25,15 @@ export default function ColorScale() {
           label="x-axis colorMap"
           value={colorX}
           onChange={(event) =>
-            setColorX(event.target.value as 'None' | 'piecewise' | 'continuous')
+            setColorX(
+              event.target.value as 'None' | 'piecewise' | 'continuous' | 'ordinal',
+            )
           }
         >
           <MenuItem value="None">None</MenuItem>
           <MenuItem value="piecewise">piecewise</MenuItem>
           <MenuItem value="continuous">continuous</MenuItem>
+          <MenuItem value="ordinal">ordinal</MenuItem>
         </TextField>
         <TextField
           select
@@ -45,15 +50,10 @@ export default function ColorScale() {
         </TextField>
       </Stack>
 
-      <LineChart
+      <BarChart
         height={300}
         grid={{ horizontal: true }}
-        series={[
-          {
-            data: [-2, -9, 12, 11, 6, -4],
-            area: true,
-          },
-        ]}
+        series={series}
         margin={{
           top: 10,
           bottom: 20,
@@ -69,25 +69,36 @@ export default function ColorScale() {
               }) ||
               (colorY === 'piecewise' && {
                 type: 'piecewise',
-                thresholds: [0, 10],
-                colors: ['red', 'green', 'blue'],
+                thresholds: [0],
+                colors: ['red', 'green'],
               }) ||
               undefined,
           },
         ]}
         xAxis={[
           {
-            scaleType: 'time',
+            scaleType: 'band',
             data: [
-              new Date(2019, 0, 1),
-              new Date(2020, 0, 1),
-              new Date(2021, 0, 1),
-              new Date(2022, 0, 1),
-              new Date(2023, 0, 1),
-              new Date(2024, 0, 1),
+              new Date(2019, 1, 1),
+              new Date(2020, 1, 1),
+              new Date(2021, 1, 1),
+              new Date(2022, 1, 1),
+              new Date(2023, 1, 1),
+              new Date(2024, 1, 1),
             ],
             valueFormatter: (value) => value.getFullYear().toString(),
             colorMap:
+              (colorX === 'ordinal' && {
+                type: 'ordinal',
+                colors: [
+                  '#ccebc5',
+                  '#a8ddb5',
+                  '#7bccc4',
+                  '#4eb3d3',
+                  '#2b8cbe',
+                  '#08589e',
+                ],
+              }) ||
               (colorX === 'continuous' && {
                 type: 'continuous',
                 min: new Date(2019, 1, 1),
@@ -109,6 +120,16 @@ export default function ColorScale() {
           '  /* ... */',
           // ColorX
           ...(colorX === 'None' ? ['  xAxis={[{}]}'] : []),
+          ...(colorX === 'ordinal'
+            ? [
+                '  xAxis={[',
+                `    {`,
+                `      type: 'ordinal',`,
+                `      colors: ['#ccebc5', '#a8ddb5', '#7bccc4', '#4eb3d3', '#2b8cbe', '#08589e']`,
+                `    }`,
+                '  ]}',
+              ]
+            : []),
           ...(colorX === 'continuous'
             ? [
                 '  xAxis={[',
@@ -137,6 +158,7 @@ export default function ColorScale() {
             ? [
                 '  yAxis={[',
                 `    {`,
+
                 `      type: 'continuous',`,
                 `      min: -10,`,
                 `      max: 10,`,
@@ -149,8 +171,8 @@ export default function ColorScale() {
             ? [
                 '  yAxis={[{',
                 `    type: 'piecewise',`,
-                `    thresholds: [0, 10],`,
-                `    colors: ['red', 'green', 'blue'],`,
+                `    thresholds: [0],`,
+                `    colors: ['red', 'green'],`,
 
                 '  }]}',
               ]
