+++ docs/data/charts/lines/lines.md
@@ -159,7 +159,7 @@ The line charts use by priority:
 
 You can find additional information about those `colorMap` properties in the [styling page](/x/react-charts/styling/#values-color).
 
-{{"demo": "ColorScale.js"}}
+{{"demo": "ColorScaleNoSnap.js"}}
 
 :::warning
 For now, ordinal config is not supported for line chart.
