+++ docs/data/charts/scatter/ColorScaleNoSnap.js
@@ -1,14 +1,19 @@
 import * as React from 'react';
-import { LineChart } from '@mui/x-charts/LineChart';
+import { ScatterChart } from '@mui/x-charts/ScatterChart';
+
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import MenuItem from '@mui/material/MenuItem';
 // @ts-ignore
 import HighlightedCode from 'docs/src/modules/components/HighlightedCode';
 
-export default function ColorScale() {
-  const [colorX, setColorX] = React.useState('None');
-  const [colorY, setColorY] = React.useState('piecewise');
+import { Chance } from 'chance';
+
+const chance = new Chance(42);
+
+export default function ColorScaleNoSnap() {
+  const [colorX, setColorX] = React.useState('piecewise');
+  const [colorY, setColorY] = React.useState('None');
 
   return (
     <Stack direction="column" spacing={1} sx={{ width: '100%', maxWidth: 600 }}>
@@ -37,59 +42,50 @@ export default function ColorScale() {
         </TextField>
       </Stack>
 
-      <LineChart
+      <ScatterChart
         height={300}
-        grid={{ horizontal: true }}
-        series={[
-          {
-            data: [-2, -9, 12, 11, 6, -4],
-            area: true,
-          },
-        ]}
+        grid={{ horizontal: true, vertical: true }}
+        series={series}
         margin={{
           top: 10,
           bottom: 20,
         }}
         yAxis={[
           {
+            min: -3,
+            max: 3,
+            tickInterval: [-3, -1.5, 0, 1.5, 3],
             colorMap:
               (colorY === 'continuous' && {
                 type: 'continuous',
-                min: -10,
-                max: 10,
-                color: ['red', 'green'],
+                min: -2,
+                max: 2,
+                color: ['blue', 'red'],
               }) ||
               (colorY === 'piecewise' && {
                 type: 'piecewise',
-                thresholds: [0, 10],
-                colors: ['red', 'green', 'blue'],
+                thresholds: [-1.5, 0, 1.5],
+                colors: ['lightblue', 'blue', 'orange', 'red'],
               }) ||
               undefined,
           },
         ]}
         xAxis={[
           {
-            scaleType: 'time',
-            data: [
-              new Date(2019, 0, 1),
-              new Date(2020, 0, 1),
-              new Date(2021, 0, 1),
-              new Date(2022, 0, 1),
-              new Date(2023, 0, 1),
-              new Date(2024, 0, 1),
-            ],
-            valueFormatter: (value) => value.getFullYear().toString(),
+            min: -3,
+            max: 3,
+            tickInterval: [-3, -1.5, 0, 1.5, 3],
             colorMap:
               (colorX === 'continuous' && {
                 type: 'continuous',
-                min: new Date(2019, 1, 1),
-                max: new Date(2024, 1, 1),
+                min: -2,
+                max: 2,
                 color: ['green', 'orange'],
               }) ||
               (colorX === 'piecewise' && {
                 type: 'piecewise',
-                thresholds: [new Date(2021, 1, 1), new Date(2023, 1, 1)],
-                colors: ['blue', 'red', 'blue'],
+                thresholds: [-1.5, 0, 1.5],
+                colors: ['#d01c8b', '#f1b6da', '#b8e186', '#4dac26'],
               }) ||
               undefined,
           },
@@ -106,8 +102,8 @@ export default function ColorScale() {
                 '  xAxis={[',
                 `    {`,
                 `      type: 'continuous',`,
-                `      min: new Date(2019, 1, 1),`,
-                `      max: new Date(2024, 1, 1),`,
+                `      min: -2,`,
+                `      max: 2,`,
                 `      color: ['green', 'orange']`,
                 `    }`,
                 '  ]}',
@@ -117,8 +113,8 @@ export default function ColorScale() {
             ? [
                 '  xAxis={[{',
                 `    type: 'piecewise',`,
-                `    thresholds: [new Date(2021, 1, 1), new Date(2023, 1, 1)],`,
-                `    colors: ['blue', 'red', 'blue'],`,
+                `    thresholds: [-1.5, 0, 1.5],`,
+                `    colors: ['#d01c8b', '#f1b6da', '#b8e186', '#4dac26'],`,
                 '  }]}',
               ]
             : []),
@@ -129,9 +125,9 @@ export default function ColorScale() {
                 '  yAxis={[',
                 `    {`,
                 `      type: 'continuous',`,
-                `      min: -10,`,
-                `      max: 10,`,
-                `      color: ['red', 'green'],`,
+                `      min: -2,`,
+                `      max: 2,`,
+                `      color: ['blue', 'red']`,
                 `    }`,
                 '  ]}',
               ]
@@ -140,8 +136,8 @@ export default function ColorScale() {
             ? [
                 '  yAxis={[{',
                 `    type: 'piecewise',`,
-                `    thresholds: [0, 10],`,
-                `    colors: ['red', 'green', 'blue'],`,
+                `    thresholds: [-1.5, 0, 1.5],`,
+                `    colors: ['lightblue', 'blue', 'orange', 'red'],`,
                 '  }]}',
               ]
             : []),
@@ -152,4 +148,25 @@ export default function ColorScale() {
       />
     </Stack>
   );
+}
+
+const series = [{ data: getGaussianSeriesData([0, 0], [1, 1], 200) }].map((s) => ({
+  ...s,
+  valueFormatter: (v) => `(${v.x.toFixed(1)}, ${v.y.toFixed(1)})`,
+}));
+
+function getGaussianSeriesData(mean, stdev = [0.3, 0.4], N = 50) {
+  return [...Array(N)].map((_, i) => {
+    const x =
+      Math.sqrt(-2.0 * Math.log(1 - chance.floating({ min: 0, max: 0.99 }))) *
+        Math.cos(2.0 * Math.PI * chance.floating({ min: 0, max: 0.99 })) *
+        stdev[0] +
+      mean[0];
+    const y =
+      Math.sqrt(-2.0 * Math.log(1 - chance.floating({ min: 0, max: 0.99 }))) *
+        Math.cos(2.0 * Math.PI * chance.floating({ min: 0, max: 0.99 })) *
+        stdev[1] +
+      mean[1];
+    return { x, y, id: i };
+  });
 }
