+++ docs/data/charts/scatter/scatter.md
@@ -66,7 +66,7 @@ The scatter charts use by priority:
 
 You can find additional information about those `colorMap` properties in the [styling page](/x/react-charts/styling/#values-color).
 
-{{"demo": "ColorScale.js"}}
+{{"demo": "ColorScaleNoSnap.js"}}
 
 ### Grid
